Self Organizing Maps
Through the use of SOMs, I plan to identify anomalies that could potentially bring new ideas to the table. The theory is as follows, anomalies stand in contrast to all others around them and therefore these papers are different and with any luck perhaps novel.

So far I have concluded that there are french articles in this corpus...lol. Obviously there is more to be done, and more cleaning that needs to take place, but at least its off to a good start.